# perf(dict): Bypass vars when possible